datapath: move fake/types.IPSet to pkg/datapath/iptables/ipset/fake.IPSet
