+++ pkg/datapath/fake/cells.go
@@ -20,6 +20,7 @@ import (
 	"github.com/cilium/cilium/pkg/datapath/iptables"
 	fakeiptables "github.com/cilium/cilium/pkg/datapath/iptables/fake"
 	"github.com/cilium/cilium/pkg/datapath/iptables/ipset"
+	fakeipset "github.com/cilium/cilium/pkg/datapath/iptables/ipset/fake"
 	"github.com/cilium/cilium/pkg/datapath/link"
 	"github.com/cilium/cilium/pkg/datapath/linux/bandwidth"
 	fakebandwidth "github.com/cilium/cilium/pkg/datapath/linux/bandwidth/fake"
@@ -76,7 +77,7 @@ var Cell = cell.Module(
 		func() lxcmap.Map { return nil },
 		func() *bigtcp.Configuration { return &bigtcp.Configuration{} },
 		func() iptables.Manager { return fakeiptables.NewManager() },
-		func() ipset.Manager { return &fakeTypes.IPSet{} },
+		func() ipset.Manager { return &fakeipset.IPSet{} },
 		func() bandwidth.Manager { return &fakebandwidth.Manager{} },
 		func() ipsec.Agent { return &fakeipsec.Agent{} },
 		func() ipsec.Config { return &fakeipsec.Config{} },
