+++ pkg/datapath/iptables/ipset/fake/ipset.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Cilium
 
-package types
+package fake
 
 import (
 	"net/netip"
